requirements.txt
Django==1.7.1
argparse==1.2.1
dj-paas-env==0.6.1
dj-static==0.0.6
django-allauth==0.18.0
djangorestframework==3.0.2
oauthlib==0.7.2
psycopg2==2.5.4
python-openid==2.2.5
requests==2.5.1
requests-oauthlib==0.4.2
static3==0.5.1
waitress==0.8.9
wsgiref==0.1.2
